move temp chart to its own class
